fix(types): kill as-any casts, complete interfaces, consolidate badges (Tier 2 PR-F)
Type safety:
- Add kind field to K8sResource base interface
- Add template.spec.containers to StatefulSet and DaemonSet interfaces
- Add initContainers to Pod interface
- Replace 3 as-any casts in ResourceDetail.tsx with typed assertions
  (Pod for logs/terminal tabs, Deployment|StatefulSet|DaemonSet for pod selector)
- Replace 2 as-any casts in resource-columns.ts with StatefulSet/DaemonSet
- Type refresh response in api.ts via APIResponse<{ accessToken: string }>

Badge consolidation:
- Extract ColorBadge to standalone components/ui/ColorBadge.tsx
- Extract SEVERITY_COLORS + SEVERITY_ORDER to lib/badge-colors.ts
- PolicyBadges.tsx: import from shared, delete local ColorBadge + SEVERITY_COLORS
- ScanBadges.tsx: import from shared, delete local SEVERITY_COLORS duplicate
- CertificateBadges.tsx: replace local CertBadge with imported ColorBadge
